+++ lib/overcommit/cli.rb
@@ -9,6 +9,8 @@ module Overcommit
       @input     = input
       @log       = logger
       @options   = {}
+
+      Overcommit::Utils.log = logger
     end
 
     def run