+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix pre-commit hook bug where amending a commit which breaks a symlink would
   result in that symlink not being included in the list of modified files
+* Fix `CaseConflicts` pre-commit hook handling of large sets of files
 
 ## 0.26.0